Moved numeric string conversion functions into own module
