+++ src/libcore/num/f64.rs
@@ -15,6 +15,7 @@ use cmp;
 use libc::{c_double, c_int};
 use libc;
 use num::NumCast;
+use num::strconv;
 use num;
 use ops;
 use option::Option;
@@ -401,8 +402,8 @@ impl num::Round for f64 {
  */
 #[inline(always)]
 pub pure fn to_str(num: f64) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 10u, true, true, num::SignNeg, num::DigAll);
+    let (r, _) = strconv::to_str_common(
+        &num, 10u, true, true, strconv::SignNeg, strconv::DigAll);
     r
 }
 
@@ -415,8 +416,8 @@ pub pure fn to_str(num: f64) -> ~str {
  */
 #[inline(always)]
 pub pure fn to_str_hex(num: f64) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 16u, true, true, num::SignNeg, num::DigAll);
+    let (r, _) = strconv::to_str_common(
+        &num, 16u, true, true, strconv::SignNeg, strconv::DigAll);
     r
 }
 
@@ -436,8 +437,8 @@ pub pure fn to_str_hex(num: f64) -> ~str {
  */
 #[inline(always)]
 pub pure fn to_str_radix(num: f64, rdx: uint) -> ~str {
-    let (r, special) = num::to_str_common(
-        &num, rdx, true, true, num::SignNeg, num::DigAll);
+    let (r, special) = strconv::to_str_common(
+        &num, rdx, true, true, strconv::SignNeg, strconv::DigAll);
     if special { fail!(~"number has a special value, \
                       try to_str_radix_special() if those are expected") }
     r
@@ -454,7 +455,8 @@ pub pure fn to_str_radix(num: f64, rdx: uint) -> ~str {
  */
 #[inline(always)]
 pub pure fn to_str_radix_special(num: f64, rdx: uint) -> (~str, bool) {
-    num::to_str_common(&num, rdx, true, true, num::SignNeg, num::DigAll)
+    strconv::to_str_common(&num, rdx, true, true,
+                           strconv::SignNeg, strconv::DigAll)
 }
 
 /**
@@ -468,8 +470,8 @@ pub pure fn to_str_radix_special(num: f64, rdx: uint) -> (~str, bool) {
  */
 #[inline(always)]
 pub pure fn to_str_exact(num: f64, dig: uint) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 10u, true, true, num::SignNeg, num::DigExact(dig));
+    let (r, _) = strconv::to_str_common(
+        &num, 10u, true, true, strconv::SignNeg, strconv::DigExact(dig));
     r
 }
 
@@ -484,8 +486,8 @@ pub pure fn to_str_exact(num: f64, dig: uint) -> ~str {
  */
 #[inline(always)]
 pub pure fn to_str_digits(num: f64, dig: uint) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 10u, true, true, num::SignNeg, num::DigMax(dig));
+    let (r, _) = strconv::to_str_common(
+        &num, 10u, true, true, strconv::SignNeg, strconv::DigMax(dig));
     r
 }
 
@@ -530,7 +532,8 @@ impl num::ToStrRadix for f64 {
  */
 #[inline(always)]
 pub pure fn from_str(num: &str) -> Option<f64> {
-    num::from_str_common(num, 10u, true, true, true, num::ExpDec, false)
+    strconv::from_str_common(num, 10u, true, true, true,
+                             strconv::ExpDec, false)
 }
 
 /**
@@ -562,7 +565,8 @@ pub pure fn from_str(num: &str) -> Option<f64> {
  */
 #[inline(always)]
 pub pure fn from_str_hex(num: &str) -> Option<f64> {
-    num::from_str_common(num, 16u, true, true, true, num::ExpBin, false)
+    strconv::from_str_common(num, 16u, true, true, true,
+                             strconv::ExpBin, false)
 }
 
 /**
@@ -586,7 +590,8 @@ pub pure fn from_str_hex(num: &str) -> Option<f64> {
  */
 #[inline(always)]
 pub pure fn from_str_radix(num: &str, rdx: uint) -> Option<f64> {
-    num::from_str_common(num, rdx, true, true, false, num::ExpNone, false)
+    strconv::from_str_common(num, rdx, true, true, false,
+                             strconv::ExpNone, false)
 }
 
 impl from_str::FromStr for f64 {
