+++ src/libcore/num/float.rs
@@ -26,6 +26,7 @@ use cmp::{Eq, Ord};
 use cmp;
 use f64;
 use num::NumCast;
+use num::strconv;
 use num;
 use ops;
 use option::{None, Option, Some};
@@ -107,8 +108,8 @@ pub mod consts {
  */
 #[inline(always)]
 pub pure fn to_str(num: float) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 10u, true, true, num::SignNeg, num::DigAll);
+    let (r, _) = strconv::to_str_common(
+        &num, 10u, true, true, strconv::SignNeg, strconv::DigAll);
     r
 }
 
@@ -121,8 +122,8 @@ pub pure fn to_str(num: float) -> ~str {
  */
 #[inline(always)]
 pub pure fn to_str_hex(num: float) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 16u, true, true, num::SignNeg, num::DigAll);
+    let (r, _) = strconv::to_str_common(
+        &num, 16u, true, true, strconv::SignNeg, strconv::DigAll);
     r
 }
 
@@ -142,8 +143,8 @@ pub pure fn to_str_hex(num: float) -> ~str {
  */
 #[inline(always)]
 pub pure fn to_str_radix(num: float, radix: uint) -> ~str {
-    let (r, special) = num::to_str_common(
-        &num, radix, true, true, num::SignNeg, num::DigAll);
+    let (r, special) = strconv::to_str_common(
+        &num, radix, true, true, strconv::SignNeg, strconv::DigAll);
     if special { fail!(~"number has a special value, \
                       try to_str_radix_special() if those are expected") }
     r
@@ -160,7 +161,8 @@ pub pure fn to_str_radix(num: float, radix: uint) -> ~str {
  */
 #[inline(always)]
 pub pure fn to_str_radix_special(num: float, radix: uint) -> (~str, bool) {
-    num::to_str_common(&num, radix, true, true, num::SignNeg, num::DigAll)
+    strconv::to_str_common(&num, radix, true, true,
+                           strconv::SignNeg, strconv::DigAll)
 }
 
 /**
@@ -174,8 +176,8 @@ pub pure fn to_str_radix_special(num: float, radix: uint) -> (~str, bool) {
  */
 #[inline(always)]
 pub pure fn to_str_exact(num: float, digits: uint) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 10u, true, true, num::SignNeg, num::DigExact(digits));
+    let (r, _) = strconv::to_str_common(
+        &num, 10u, true, true, strconv::SignNeg, strconv::DigExact(digits));
     r
 }
 
@@ -196,8 +198,8 @@ pub fn test_to_str_exact_do_decimal() {
  */
 #[inline(always)]
 pub pure fn to_str_digits(num: float, digits: uint) -> ~str {
-    let (r, _) = num::to_str_common(
-        &num, 10u, true, true, num::SignNeg, num::DigMax(digits));
+    let (r, _) = strconv::to_str_common(
+        &num, 10u, true, true, strconv::SignNeg, strconv::DigMax(digits));
     r
 }
 
@@ -242,7 +244,8 @@ impl num::ToStrRadix for float {
  */
 #[inline(always)]
 pub pure fn from_str(num: &str) -> Option<float> {
-    num::from_str_common(num, 10u, true, true, true, num::ExpDec, false)
+    strconv::from_str_common(num, 10u, true, true, true,
+                             strconv::ExpDec, false)
 }
 
 /**
@@ -274,7 +277,8 @@ pub pure fn from_str(num: &str) -> Option<float> {
  */
 #[inline(always)]
 pub pure fn from_str_hex(num: &str) -> Option<float> {
-    num::from_str_common(num, 16u, true, true, true, num::ExpBin, false)
+    strconv::from_str_common(num, 16u, true, true, true,
+                             strconv::ExpBin, false)
 }
 
 /**
@@ -298,7 +302,8 @@ pub pure fn from_str_hex(num: &str) -> Option<float> {
  */
 #[inline(always)]
 pub pure fn from_str_radix(num: &str, radix: uint) -> Option<float> {
-    num::from_str_common(num, radix, true, true, false, num::ExpNone, false)
+    strconv::from_str_common(num, radix, true, true, false,
+                             strconv::ExpNone, false)
 }
 
 impl from_str::FromStr for float {
